Revert "merge all commit to my forked repo (#1)"
This reverts commit 1de128fc40c7e4190279ba7394e8437c220c7e91.

+++ hbase-metrics/src/main/java/org/apache/hadoop/hbase/metrics/impl/HistogramImpl.java
@@ -36,7 +36,7 @@ public class HistogramImpl implements Histogram {
   private final CounterImpl counter;
 
   public HistogramImpl() {
-    this((long) Integer.MAX_VALUE << 2);
+    this(Integer.MAX_VALUE << 2);
   }
 
   public HistogramImpl(long maxExpected) {
